refactor(Validator): 파일 경로 수정
domain/ 디렉터리 하위로 이동
